Extract data loading to module
